last commit before I quit

+++ remote-dashboard/src/app/servers/servers.component.ts
@@ -52,31 +52,83 @@ export class ServersComponent implements OnInit {
       if (ServerToDelete.localServerId == server.localServerId) {
 
         try {
-          await this.http.delete<LocalServer>(this.servers_url + "/" + ServerToDelete.localServerId).toPromise();
+          swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            type: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then(async (result) => {
+            if (result.value) {
+              await this.http.delete<LocalServer>(this.servers_url + "/" + ServerToDelete.localServerId).toPromise();
+              swal.fire(
+                'Deleted!',
+                'The server has been deleted.',
+                'success'
+              )
+            }
+          })          
         }
         catch (error) { }
       }
     });
   }
 
+  async postNewServer() {
 
+    
+      swal.mixin({
+        input: 'text',
+        confirmButtonText: 'Next &rarr;',
+        showCancelButton: true,
+        progressSteps: ['name', 'Mac Address', 'Users']
+      }).queue([
+        {
+          title: 'Enter Name',
+          text: this.NewServer.displayName
+        },
+        {
+          title: 'Enter Mac Address',
+          text: this.NewServer.macAddress
+        },
+        {
+          title: 'Enter Users',
+          text: this.NewServer.validUsers[0]
+        }
+      ]).then(async (result) => {
+        if (result.value) {
+          try{
+          await this.http.post(this.servers_url, {
+            displayName: result.value[0],
+            localServerId: "",
+            macAddress: result.value[1],
+            validUsers: result.value[2],
+            connectionStatus: false
+          }).toPromise(); 
+          this.servers.push(this.NewServer);
+        this.NewServer = null;
+          swal.fire({
+            title: 'All done!',
+            html:
+              'Your Server: <pre><code>' +
+                JSON.stringify(result.value) +
+              '</code></pre>',
+            confirmButtonText: 'Lovely!'
+        
+          })
+        }
+        catch (error) {
+          console.log(error);
+        }
 
-  async postNewServer(NewServer: LocalServer) {
-
-    try {
-      await this.http.post(this.servers_url, {
-        displayName: NewServer.displayName,
-        localServerId: NewServer.localServerId,
-        macAddress: NewServer.macAddress,
-        validUsers: NewServer.validUsers,
-        connectionStatus: NewServer.connectionStatus
-      }).toPromise();
-      this.servers.push(NewServer);
     }
-    catch (error) {
-      console.log(error);
-    }
-  }
+
+   }
+      )}
+   
+       
 
 
   async editServer(server: LocalServer) {
@@ -93,19 +145,31 @@ export class ServersComponent implements OnInit {
         }
       }
     });
-
   }
-
+  
   async authKeyRequest(server: LocalServer) {
     this.servers.forEach(async serv => {
 
       if (serv.localServerId == server.localServerId) {
 
         try {
-          const apikey = await this.http.post<string>(`${this.servers_url}/auth/${server.localServerId}`, {}).toPromise();
-          swal.fire({type: 'info', title: 'key for authentication' , showCloseButton: true, html:   
-          apikey + '<br/><br/>'  + 
-           '<b> CAUTION: you may be disconnected now  </b>'})
+          swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            type: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, generate key!'
+          }).then(async (result) => {
+            if (result.value) {
+              const apikey = await this.http.post<string>(`${this.servers_url}/auth/${server.localServerId}`, {}).toPromise();
+              swal.fire({type: 'info', title: 'key for authentication' , showCloseButton: true, html:   
+              apikey + '<br/><br/>'  + 
+               '<b> CAUTION: you may be disconnected now  </b>'});
+            }
+          })
+
            
         }
         catch (error) {
